UI is working now.
CONSOLE button needs to be fixed to proper IP address
installing animations needs to be added

+++ appscale-plugin/plugin-appscale-impl/src/main/java/io/subutai/plugin/appscale/impl/ClusterConfiguration.java
@@ -6,8 +6,6 @@
 package io.subutai.plugin.appscale.impl;
 
 
-import java.util.logging.Level;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,10 +20,10 @@ import io.subutai.common.peer.LocalPeer;
 import io.subutai.common.peer.ResourceHost;
 import io.subutai.common.tracker.TrackerOperation;
 import io.subutai.core.peer.api.PeerManager;
-import io.subutai.plugin.appscale.api.AppScaleConfig;
 import io.subutai.core.plugincommon.api.ClusterConfigurationException;
 import io.subutai.core.plugincommon.api.ClusterConfigurationInterface;
 import io.subutai.core.plugincommon.api.ConfigBase;
+import io.subutai.plugin.appscale.api.AppScaleConfig;
 
 
 /**
@@ -72,20 +70,23 @@ public class ClusterConfiguration implements ClusterConfigurationInterface
         try
         {
             containerHost = environment.getContainerHostByHostname ( config.getClusterName () );
+
+            LOG.info (
+                    "Container Host Found: " + containerHost.getContainerId () + "\n"
+                    + "\n" + containerHost.getHostname () + "\n" );
         }
         catch ( ContainerHostNotFoundException ex )
         {
             LOG.error ( "configureCluster " + ex );
+            // po.addLogFailed( "container host is not found : " + ex );
         }
-        LOG.info (
-                "Container Host Found: " + containerHost.getContainerId () + "\n"
-                + "\n" + containerHost.getHostname () + "\n" );
+
 
         // this.commandExecute ( containerHost, Commands.getRemoveSubutaiList () );
         this.commandExecute ( containerHost, Commands.getCreateLogDir () );
 
         LOG.info ( "installing appscale can take several minutes." );
-        po.addLog ( "installing appscale can take several minutes." );
+
 
         // AppScalefile configuration
         this.appscaleInitCluster ( containerHost, environment, config ); // writes AppScalefile
@@ -99,7 +100,7 @@ public class ClusterConfiguration implements ClusterConfigurationInterface
         LOG.info ( "Run shell starting..." );
         this.commandExecute ( containerHost, Commands.getRunShell () );
         LOG.info ( "Run shell completed..." );
-        LOG.info ( "Correcting Hostname" );
+        // LOG.info ( "Correcting Hostname" );
         // this.commandExecute ( containerHost, "sudo echo '" + config.getClusterName () + "' > /etc/hostname" );
 
 
@@ -107,17 +108,17 @@ public class ClusterConfiguration implements ClusterConfigurationInterface
                 + "ssh -f -N -R 1443:<Your RH IP>:1443 ubuntu@localhost\n"
                 + "ssh -f -N -R 5555:<Your RH IP>:5555 ubuntu@localhost\n"
                 + "ssh -f -N -R 8081:<Your RH IP>:8081 ubuntu@localhost\n" );
-        po.addLog ( "Login into your RH and run these commands: \n"
-                + "ssh -f -N -R 1443:<Your RH IP>:1443 ubuntu@localhost\n"
-                + "ssh -f -N -R 5555:<Your RH IP>:5555 ubuntu@localhost\n"
-                + "ssh -f -N -R 8081:<Your RH IP>:8081 ubuntu@localhost\n" );
-
+        /*
+         * po.addLog ( "Login into your RH and run these commands: \n" + "ssh -f -N -R 1443:<Your RH IP>:1443
+         * ubuntu@localhost\n" + "ssh -f -N -R 5555:<Your RH IP>:5555 ubuntu@localhost\n" + "ssh -f -N -R 8081:<Your RH
+         * IP>:8081 ubuntu@localhost\n" );
+         */
         config.setEnvironmentId ( environment.getId () );
         appscaleManager.getPluginDAO ().saveInfo ( AppScaleConfig.PRODUCT_KEY, configBase.getClusterName (),
                                                    configBase );
         LOG.info ( "Appscale saved to database" );
-        po.addLogDone ( "Appscale is saved to database" );
-
+        po.addLogDone ( "Appscale is saved to database" ); // will try to last po.addLogDone... nothing else..
+        this.runShhInResourceHost ( containerHost );
 
     }
 
@@ -129,17 +130,20 @@ public class ClusterConfiguration implements ClusterConfigurationInterface
         String ipAddress = this.getIPAddress ( containerHost ); // only for master
         String command1443 = "ssh -f -N -R 1443:" + ipAddress + ":1443 ubuntu@localhost";
         String command5555 = "ssh -f -N -R 5555:" + ipAddress + ":5555 ubuntu@localhost";
+        String command8080 = "ssh -f -N -R 8081:" + ipAddress + ":8080 ubuntu@localhost";
         String command8081 = "ssh -f -N -R 8081:" + ipAddress + ":8081 ubuntu@localhost";
         try
         {
             ResourceHost resourceHostByContainerId = localPeer.getResourceHostByContainerId ( containerHost.getId () );
+            LOG.info ( "HERE IS RESOURCE HOST: " + resourceHostByContainerId.getHostname () );
             resourceHostByContainerId.execute ( new RequestBuilder ( command1443 ) );
             resourceHostByContainerId.execute ( new RequestBuilder ( command5555 ) );
+            resourceHostByContainerId.execute ( new RequestBuilder ( command8080 ) );
             resourceHostByContainerId.execute ( new RequestBuilder ( command8081 ) );
         }
         catch ( HostNotFoundException | CommandException ex )
         {
-            java.util.logging.Logger.getLogger ( ClusterConfiguration.class.getName () ).log ( Level.SEVERE, null, ex );
+            LOG.error ( ex.toString () );
         }
 
     }
@@ -147,9 +151,27 @@ public class ClusterConfiguration implements ClusterConfigurationInterface
 
     private void makeCleanUpPreviousInstallation ( EnvironmentContainerHost containerHost )
     {
-        this.commandExecute ( containerHost, "sudo rm /root/.ssh/*" );
-        this.commandExecute ( containerHost, "sudo touch /root/.ssh/known_hosts" );
-        this.commandExecute ( containerHost, "sudo rm /root/.appscale/*" );
+        try
+        {
+            CommandResult cr;
+            cr = containerHost.execute ( new RequestBuilder ( "sudo ls /root/.ssh" ) );
+            if ( !cr.toString ().equals ( "" ) )
+            {
+                this.commandExecute ( containerHost, "sudo rm /root/.ssh/*" );
+                this.commandExecute ( containerHost, "sudo touch /root/.ssh/known_hosts" );
+            }
+            cr = containerHost.execute ( new RequestBuilder ( "sudo ls /root/.appscale" ) );
+            if ( !cr.toString ().equals ( "" ) )
+            {
+                this.commandExecute ( containerHost, "sudo rm /root/.appscale/*" );
+            }
+
+        }
+        catch ( CommandException ex )
+        {
+            LOG.error ( " clean process command exception: " + ex );
+            // po.addLogFailed( "Clean up failed..." );
+        }
     }
 
 
@@ -159,11 +181,12 @@ public class ClusterConfiguration implements ClusterConfigurationInterface
         {
             // downloading takes time...
             CommandResult responseFrom = containerHost.execute ( new RequestBuilder ( command ).withTimeout ( 8000 ) );
-            po.addLogDone ( command + " executed with: " + responseFrom );
+
         }
         catch ( CommandException e )
         {
             LOG.error ( "Error while executing \"" + command + "\".\n" + e );
+            // po.addLogFailed( command + " can not be executed properly..." );
         }
     }
 
@@ -202,7 +225,7 @@ public class ClusterConfiguration implements ClusterConfigurationInterface
             }
             catch ( ContainerHostNotFoundException ex )
             {
-
+                LOG.error ( ex.toString () );
             }
         }
         else
